review: split channel membership commands into separate files
Per @scottlovegrove's review feedback. Each subcommand now lives in its
own file, mirroring the existing channel/ folder pattern (list.ts,
threads.ts, helpers.ts):

- src/commands/channel/membership-helpers.ts: shared bits used by add,
  remove, sync, and the members list view (channelUserIds,
  fetchUsersByIds, logExpansion, describeExpansion, groupsFullyInChannel,
  mutateChannelMembership).
- src/commands/channel/members.ts: now only the list handler.
- src/commands/channel/add.ts, remove.ts: thin wrappers around
  mutateChannelMembership.
- src/commands/channel/sync.ts: the sync handler (kept here because of
  the self-removal guard and dry-run-by-default behaviour).
- src/commands/channel/index.ts: imports updated.

Tests split to match:

- members.test.ts: only the list tests.
- add.test.ts, remove.test.ts, sync.test.ts: one file per command, each
  with its own minimal mock setup.

677 tests still passing (43 test files, +3 from the split).
