+++ src/commands/channel/add.ts
@@ -1,0 +1,9 @@
+import { type ChannelMutationOptions, mutateChannelMembership } from './membership-helpers.js'
+
+export async function addChannelMembers(
+    channelRef: string,
+    refs: string[],
+    options: ChannelMutationOptions,
+): Promise<void> {
+    return mutateChannelMembership(channelRef, refs, 'add', options)
+}
