+++ src/commands/channel/members.test.ts
@@ -68,11 +68,9 @@ beforeEach(() => {
         channels: { getChannel: mockGetChannel },
         batch: mockBatch,
     })
-    // The default real-world behaviour: pass through response.data, ignoring the label.
     apiMocks.getOptionalBatchData.mockImplementation((response: { code: number; data: unknown }) =>
         response && response.code < 400 ? (response.data ?? null) : null,
     )
-    apiMocks.getSessionUser.mockResolvedValue({ id: 1, name: 'Me', email: 'me@d.com' })
     apiMocks.getWorkspaceGroups.mockResolvedValue([frontendGroup, backendGroup, allInChannel])
     refsMocks.resolveChannelRef.mockResolvedValue(sampleChannel)
 })
@@ -172,368 +170,5 @@ describe('tw channel members (list)', () => {
         const output = JSON.parse(line)
         expect(output).toHaveProperty('creator')
         expect(output).toHaveProperty('version')
-    })
-})
-
-describe('tw channel add', () => {
-    it('adds users (no group expansion)', async () => {
-        refsMocks.resolveChannelMemberRefs.mockResolvedValue({
-            userIds: [4, 5],
-            expandedFrom: [],
-        })
-        const program = createProgram()
-        vi.spyOn(console, 'log').mockImplementation(() => {})
-
-        await program.parseAsync([
-            'node',
-            'tw',
-            'channel',
-            'add',
-            'general',
-            'dave@d.com',
-            'eve@d.com',
-        ])
-
-        expect(refsMocks.resolveChannelMemberRefs).toHaveBeenCalledWith(
-            ['dave@d.com', 'eve@d.com'],
-            1,
-        )
-        expect(apiMocks.addUsersToChannel).toHaveBeenCalledWith(500, [4, 5])
-    })
-
-    it('expands group:<ref> and dedupes', async () => {
-        refsMocks.resolveChannelMemberRefs.mockResolvedValue({
-            userIds: [4, 5, 6],
-            expandedFrom: [{ groupId: 200, groupName: 'Backend', userIds: [4, 5] }],
-        })
-        const program = createProgram()
-        const consoleSpy = vi.spyOn(console, 'log').mockImplementation(() => {})
-
-        await program.parseAsync([
-            'node',
-            'tw',
-            'channel',
-            'add',
-            'general',
-            'group:Backend',
-            'frank',
-        ])
-
-        expect(apiMocks.addUsersToChannel).toHaveBeenCalledWith(500, [4, 5, 6])
-        const text = consoleSpy.mock.calls.map((c) => c[0]).join('\n')
-        expect(text).toContain('Backend')
-    })
-
-    it('skips users already in channel', async () => {
-        // channel has [1,2,3]; requesting [2,4] → only 4 should be added, 2 skipped
-        refsMocks.resolveChannelMemberRefs.mockResolvedValue({
-            userIds: [2, 4],
-            expandedFrom: [],
-        })
-        const program = createProgram()
-        vi.spyOn(console, 'log').mockImplementation(() => {})
-
-        await program.parseAsync(['node', 'tw', 'channel', 'add', 'general', 'id:2', 'id:4'])
-
-        expect(apiMocks.addUsersToChannel).toHaveBeenCalledWith(500, [4])
-    })
-
-    it('makes no API call when all users already members', async () => {
-        refsMocks.resolveChannelMemberRefs.mockResolvedValue({
-            userIds: [1, 2],
-            expandedFrom: [],
-        })
-        const program = createProgram()
-        vi.spyOn(console, 'log').mockImplementation(() => {})
-
-        await program.parseAsync(['node', 'tw', 'channel', 'add', 'general', 'id:1', 'id:2'])
-
-        expect(apiMocks.addUsersToChannel).not.toHaveBeenCalled()
-    })
-
-    it('--dry-run does not mutate', async () => {
-        refsMocks.resolveChannelMemberRefs.mockResolvedValue({
-            userIds: [4, 5],
-            expandedFrom: [],
-        })
-        const program = createProgram()
-        const consoleSpy = vi.spyOn(console, 'log').mockImplementation(() => {})
-
-        await program.parseAsync([
-            'node',
-            'tw',
-            'channel',
-            'add',
-            'general',
-            'id:4',
-            'id:5',
-            '--dry-run',
-        ])
-
-        expect(apiMocks.addUsersToChannel).not.toHaveBeenCalled()
-        expect(consoleSpy.mock.calls.some((c) => String(c[0]).includes('[dry-run]'))).toBe(true)
-    })
-
-    it('--json shape includes added/alreadyMembers', async () => {
-        refsMocks.resolveChannelMemberRefs.mockResolvedValue({
-            userIds: [2, 4],
-            expandedFrom: [],
-        })
-        const program = createProgram()
-        const consoleSpy = vi.spyOn(console, 'log').mockImplementation(() => {})
-
-        await program.parseAsync([
-            'node',
-            'tw',
-            'channel',
-            'add',
-            'general',
-            'id:2',
-            'id:4',
-            '--json',
-        ])
-
-        const output = JSON.parse(consoleSpy.mock.calls[0][0])
-        expect(output).toMatchObject({ id: 500, added: [4], alreadyMembers: [2] })
-    })
-})
-
-describe('tw channel remove', () => {
-    it('only removes users that are members', async () => {
-        refsMocks.resolveChannelMemberRefs.mockResolvedValue({
-            userIds: [2, 3, 99],
-            expandedFrom: [],
-        })
-        const program = createProgram()
-        vi.spyOn(console, 'log').mockImplementation(() => {})
-
-        await program.parseAsync([
-            'node',
-            'tw',
-            'channel',
-            'remove',
-            'general',
-            'id:2',
-            'id:3',
-            'id:99',
-        ])
-
-        expect(apiMocks.removeUsersFromChannel).toHaveBeenCalledWith(500, [2, 3])
-    })
-
-    it('makes no API call when none of the users are members', async () => {
-        refsMocks.resolveChannelMemberRefs.mockResolvedValue({
-            userIds: [99, 100],
-            expandedFrom: [],
-        })
-        const program = createProgram()
-        vi.spyOn(console, 'log').mockImplementation(() => {})
-
-        await program.parseAsync(['node', 'tw', 'channel', 'remove', 'general', 'id:99', 'id:100'])
-
-        expect(apiMocks.removeUsersFromChannel).not.toHaveBeenCalled()
-    })
-
-    it('--dry-run does not mutate', async () => {
-        refsMocks.resolveChannelMemberRefs.mockResolvedValue({
-            userIds: [2, 3],
-            expandedFrom: [],
-        })
-        const program = createProgram()
-        const consoleSpy = vi.spyOn(console, 'log').mockImplementation(() => {})
-
-        await program.parseAsync([
-            'node',
-            'tw',
-            'channel',
-            'remove',
-            'general',
-            'id:2',
-            'id:3',
-            '--dry-run',
-        ])
-
-        expect(apiMocks.removeUsersFromChannel).not.toHaveBeenCalled()
-        expect(consoleSpy.mock.calls.some((c) => String(c[0]).includes('[dry-run]'))).toBe(true)
-    })
-
-    it('--json shape includes removed/notMembers', async () => {
-        refsMocks.resolveChannelMemberRefs.mockResolvedValue({
-            userIds: [2, 99],
-            expandedFrom: [],
-        })
-        const program = createProgram()
-        const consoleSpy = vi.spyOn(console, 'log').mockImplementation(() => {})
-
-        await program.parseAsync([
-            'node',
-            'tw',
-            'channel',
-            'remove',
-            'general',
-            'id:2',
-            'id:99',
-            '--json',
-        ])
-
-        const output = JSON.parse(consoleSpy.mock.calls[0][0])
-        expect(output).toMatchObject({ id: 500, removed: [2], notMembers: [99] })
-    })
-
-    it('--full fetches and prints the updated channel', async () => {
-        refsMocks.resolveChannelMemberRefs.mockResolvedValue({
-            userIds: [2],
-            expandedFrom: [],
-        })
-        mockGetChannel.mockResolvedValue({ ...sampleChannel, userIds: [1, 3] })
-        const program = createProgram()
-        const consoleSpy = vi.spyOn(console, 'log').mockImplementation(() => {})
-
-        await program.parseAsync([
-            'node',
-            'tw',
-            'channel',
-            'remove',
-            'general',
-            'id:2',
-            '--json',
-            '--full',
-        ])
-
-        expect(mockGetChannel).toHaveBeenCalledWith(500)
-        const output = JSON.parse(consoleSpy.mock.calls[0][0])
-        // --full returns the full SDK channel object
-        expect(output.id).toBe(500)
-        expect(output).toHaveProperty('creator')
-        expect(output).toHaveProperty('version')
-    })
-})
-
-describe('tw channel sync', () => {
-    it('dry-run by default — does not mutate', async () => {
-        refsMocks.resolveChannelMemberRefs.mockResolvedValue({
-            userIds: [1, 4, 5],
-            expandedFrom: [],
-        })
-        const program = createProgram()
-        const consoleSpy = vi.spyOn(console, 'log').mockImplementation(() => {})
-
-        await program.parseAsync([
-            'node',
-            'tw',
-            'channel',
-            'sync',
-            'general',
-            'id:1',
-            'id:4',
-            'id:5',
-        ])
-
-        expect(apiMocks.addUsersToChannel).not.toHaveBeenCalled()
-        expect(apiMocks.removeUsersFromChannel).not.toHaveBeenCalled()
-        const text = consoleSpy.mock.calls.map((c) => c[0]).join('\n')
-        expect(text).toContain('[dry-run]')
-    })
-
-    it('--apply computes add/remove diff', async () => {
-        // channel currently [1,2,3]; desired [1,4,5] → add [4,5], remove [2,3]
-        refsMocks.resolveChannelMemberRefs.mockResolvedValue({
-            userIds: [1, 4, 5],
-            expandedFrom: [],
-        })
-        const program = createProgram()
-        vi.spyOn(console, 'log').mockImplementation(() => {})
-
-        await program.parseAsync([
-            'node',
-            'tw',
-            'channel',
-            'sync',
-            'general',
-            'id:1',
-            'id:4',
-            'id:5',
-            '--apply',
-        ])
-
-        expect(apiMocks.addUsersToChannel).toHaveBeenCalledWith(500, [4, 5])
-        expect(apiMocks.removeUsersFromChannel).toHaveBeenCalledWith(500, [2, 3])
-    })
-
-    it('refuses to remove acting user without --include-self', async () => {
-        // session user id:1 is in current channel; desired set omits id:1 → would remove self
-        refsMocks.resolveChannelMemberRefs.mockResolvedValue({
-            userIds: [2, 3, 4],
-            expandedFrom: [],
-        })
-        const program = createProgram()
-
-        await expect(
-            program.parseAsync([
-                'node',
-                'tw',
-                'channel',
-                'sync',
-                'general',
-                'id:2',
-                'id:3',
-                'id:4',
-                '--apply',
-            ]),
-        ).rejects.toMatchObject({ code: 'INVALID_VALUE' })
-    })
-
-    it('--include-self allows removing acting user', async () => {
-        refsMocks.resolveChannelMemberRefs.mockResolvedValue({
-            userIds: [2, 3, 4],
-            expandedFrom: [],
-        })
-        const program = createProgram()
-        vi.spyOn(console, 'log').mockImplementation(() => {})
-
-        await program.parseAsync([
-            'node',
-            'tw',
-            'channel',
-            'sync',
-            'general',
-            'id:2',
-            'id:3',
-            'id:4',
-            '--apply',
-            '--include-self',
-        ])
-
-        expect(apiMocks.addUsersToChannel).toHaveBeenCalledWith(500, [4])
-        expect(apiMocks.removeUsersFromChannel).toHaveBeenCalledWith(500, [1])
-    })
-
-    it('--json output shape', async () => {
-        refsMocks.resolveChannelMemberRefs.mockResolvedValue({
-            userIds: [1, 4],
-            expandedFrom: [{ groupId: 200, groupName: 'Backend', userIds: [4] }],
-        })
-        const program = createProgram()
-        const consoleSpy = vi.spyOn(console, 'log').mockImplementation(() => {})
-
-        await program.parseAsync([
-            'node',
-            'tw',
-            'channel',
-            'sync',
-            'general',
-            'id:1',
-            'group:Backend',
-            '--apply',
-            '--json',
-        ])
-
-        const output = JSON.parse(consoleSpy.mock.calls[0][0])
-        expect(output).toMatchObject({
-            id: 500,
-            added: [4],
-            removed: [2, 3],
-            expandedFrom: [{ groupId: 200, groupName: 'Backend', userIds: [4] }],
-        })
     })
 })
